pyproject.toml
[tool.poetry]
name = "private-bin"
version = "0.1.0"
description = ""
authors = ["MykA012 <aleshin.tomsk2@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.14"
SQLAlchemy = "^2.0.41"
uvicorn = "^0.34.3"
pydantic-settings = "^2.10.1"
alembic = "^1.16.2"
asyncpg = "^0.30.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
